lmb-1257: restructure to controller and service
